Remove types directory, regenerate protobufs, and prepare for sharing the KVStore.
