+++ modules/light-clients/10-wasm/errors.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
