+++ modules/light-clients/10-wasm/fail_kv_store.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"io"
