+++ modules/light-clients/10-wasm/header.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"fmt"
@@ -10,10 +10,6 @@ var _ exported.ClientMessage = &Header{}
 
 func (m Header) ClientType() string {
 	return exported.Wasm
-}
-
-func (m Header) GetHeight() exported.Height {
-	return m.Height
 }
 
 func (m Header) ValidateBasic() error {
