+++ modules/light-clients/10-wasm/keys.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"encoding/hex"
