+++ modules/light-clients/10-wasm/misbehavior.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"github.com/cosmos/ibc-go/modules/core/exported"
@@ -8,15 +8,15 @@ var (
 	_ exported.Misbehaviour = &Misbehaviour{}
 )
 
-func (m *Misbehaviour) ClientType() string {
+func (m Misbehaviour) ClientType() string {
 	return m.Header1.ClientType()
 }
 
-func (m *Misbehaviour) GetClientID() string {
+func (m Misbehaviour) GetClientID() string {
 	return m.ClientId
 }
 
-func (m *Misbehaviour) ValidateBasic() error {
+func (m Misbehaviour) ValidateBasic() error {
 	if err := m.Header1.ValidateBasic(); err != nil {
 		return err
 	}
