+++ modules/light-clients/10-wasm/store.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"bytes"
