+++ modules/light-clients/10-wasm/validation.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import cosmwasm "github.com/CosmWasm/wasmvm"
