+++ modules/light-clients/10-wasm/vm.go
@@ -1,4 +1,4 @@
-package types
+package wasm
 
 import (
 	"bytes"
@@ -94,19 +94,19 @@ func CreateVM(vmConfig *VMConfig, validationConfig *ValidationConfig) {
 	WasmVal = wasmValidator
 }
 
-func PushNewWasmCode(store sdk.KVStore, c *ClientState, code []byte) error {
+func PushNewWasmCode(store sdk.KVStore, code []byte) (cosmwasm.Checksum, error) {
 	// check to see if the store has a code with the same code id
 	codeHash := generateWasmCodeHash(code)
 	codeIDKey := CodeID(codeHash)
 	if store.Has(codeIDKey) {
-		return ErrWasmCodeExists
+		return nil, ErrWasmCodeExists
 	}
 
 	// run the code through the wasm light client validation process
 	if isValidWasmCode, err := WasmVal.validateWasmCode(code); err != nil {
-		return sdkerrors.Wrapf(ErrWasmCodeValidation, "unable to validate wasm code: %s", err)
+		return nil, sdkerrors.Wrapf(ErrWasmCodeValidation, "unable to validate wasm code: %s", err)
 	} else if !isValidWasmCode {
-		return ErrWasmInvalidCode
+		return nil, ErrWasmInvalidCode
 	}
 
 	// create the code in the vm
@@ -114,22 +114,21 @@ func PushNewWasmCode(store sdk.KVStore, c *ClientState, code []byte) error {
 	// is no code with the same hash?
 	codeID, err := WasmVM.Create(code)
 	if err != nil {
-		return ErrWasmInvalidCode
+		return nil, ErrWasmInvalidCode
 	}
 
 	// safety check to assert that code id returned by WasmVM equals to code hash
 	if !bytes.Equal(codeID, codeHash) {
-		return ErrWasmInvalidCodeID
+		return nil, ErrWasmInvalidCodeID
 	}
 
 	store.Set(codeIDKey, code)
-	c.CodeId = codeID
-	return nil
+	return codeID, nil
 }
 
 // Calls vm.Init with appropriate arguments
 // TODO: Move this into a public method on the 28-wasm keeper
-func initContract(codeID []byte, ctx sdk.Context, store sdk.KVStore, msg []byte) (*types.Response, error) {
+func initContract(codeID []byte, ctx sdk.Context, store sdk.KVStore) (*types.Response, error) {
 	gasMeter := ctx.GasMeter()
 	chainID := ctx.BlockHeader().ChainID
 	height := ctx.BlockHeader().Height
@@ -160,7 +159,7 @@ func initContract(codeID []byte, ctx sdk.Context, store sdk.KVStore, msg []byte)
 	// mockQuerier := api.MockQuerier{}
 
 	desercost := types.UFraction{Numerator: 0, Denominator: 1}
-	response, _, err := WasmVM.Instantiate(codeID, env, msgInfo, msg, store, cosmwasm.GoAPI{}, nil, gasMeter, gasMeter.Limit(), desercost)
+	response, _, err := WasmVM.Instantiate(codeID, env, msgInfo, []byte("{}"), store, cosmwasm.GoAPI{}, nil, gasMeter, gasMeter.Limit(), desercost)
 	return response, err
 }
 
@@ -189,12 +188,12 @@ func callContract(codeID []byte, ctx sdk.Context, store sdk.KVStore, msg []byte)
 		},
 	}
 
-	return callContractWithEnvAndMeter(codeID, &ctx, store, env, gasMeter, msg)
+	return callContractWithEnvAndMeter(codeID, ctx, store, env, gasMeter, msg)
 }
 
 // Calls vm.Execute with supplied environment and gas meter
 // TODO: Move this into a private method on the 28-wasm keeper
-func callContractWithEnvAndMeter(codeID cosmwasm.Checksum, ctx *sdk.Context, store cosmwasm.KVStore, env types.Env, gasMeter sdk.GasMeter, msg []byte) (*types.Response, error) {
+func callContractWithEnvAndMeter(codeID cosmwasm.Checksum, ctx sdk.Context, store sdk.KVStore, env types.Env, gasMeter sdk.GasMeter, msg []byte) (*types.Response, error) {
 	msgInfo := types.MessageInfo{
 		Sender: "",
 		Funds:  nil,
@@ -203,15 +202,15 @@ func callContractWithEnvAndMeter(codeID cosmwasm.Checksum, ctx *sdk.Context, sto
 	// mockFailureAPI := *api.NewMockFailureAPI()
 	// mockQuerier := api.MockQuerier{}
 	desercost := types.UFraction{Numerator: 1, Denominator: 1}
-	resp, gasUsed, err := WasmVM.Execute(codeID, env, msgInfo, msg, store, cosmwasm.GoAPI{}, nil, nil, gasMeter.Limit(), desercost)
-	if ctx != nil {
-		consumeGas(*ctx, gasUsed)
+	resp, gasUsed, err := WasmVM.Execute(codeID, env, msgInfo, msg, store, cosmwasm.GoAPI{}, nil, gasMeter, gasMeter.Limit(), desercost)
+	if &ctx != nil {
+		consumeGas(ctx, gasUsed)
 	}
 	return resp, err
 }
 
 // TODO: Move this into a public method on the 28-wasm keeper
-func queryContractWithStore(codeID cosmwasm.Checksum, store cosmwasm.KVStore, msg []byte) ([]byte, error) {
+func queryContractWithStore(codeID cosmwasm.Checksum, store sdk.KVStore, msg []byte) ([]byte, error) {
 	// TODO: fix this
 	// mockEnv := api.MockEnv()
 	// mockGasMeter := api.NewMockGasMeter(1)
