+++ testing/chain.go
@@ -31,7 +31,7 @@ import (
 	"github.com/cosmos/ibc-go/v5/modules/core/exported"
 	"github.com/cosmos/ibc-go/v5/modules/core/types"
 	ibctm "github.com/cosmos/ibc-go/v5/modules/light-clients/07-tendermint"
-	wasm "github.com/cosmos/ibc-go/v5/modules/light-clients/10-wasm/types"
+	wasm "github.com/cosmos/ibc-go/v5/modules/light-clients/10-wasm"
 	"github.com/cosmos/ibc-go/v5/testing/mock"
 	"github.com/cosmos/ibc-go/v5/testing/simapp"
 )
